# docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: A robust microservice backend for managing teams, users, and digital
    assets.
  termsOfService: http://swagger.io/terms/
  title: Team Management API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: 'Example: {''email'': ''t@t.com'', ''password'': ''123''}'
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login a user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with a specified role (manager, member)
      parameters:
      - description: 'Example: {''username'': ''test'', ''email'': ''t@t.com'', ''password'':
          ''123'', ''role'': ''manager''}'
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"